# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Local Grocery Marketplace API
  description: API specification for hyperlocal grocery marketplace (India)
  version: 1.0.0
servers:
  - url: https://api.yourdomain.com/v1
    description: Production
  - url: https://staging-api.yourdomain.com/v1
    description: Staging
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
        imageUrl:
          type: string
...
security:
  - BearerAuth: []
paths:
...
  /stores:
    post:
      tags: [Retailer]
      summary: Create store
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
      responses:
        '201':
          description: Store created

  ## CATALOG
  /catalog/products:
    get:
      tags: [Catalog]
      summary: List products
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /catalog/products/{id}:
    get:
      tags: [Catalog]
      summary: Product details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  ## CART
...